add tests for 'Tymczak' and 'Lukasiewicz'
The example string 'Tymczak' is used in the wikipedia article for
Soundex to illustrate the difference between the original soundex
algorithm and what is followed by most SQL languages.

It also shows a failure in the mapping table for 'k' and 'z'.

I've also added a failing test for 'Lukasiewicz'.